requirements.txt
# Core ML
numpy
pandas
scikit-learn
xgboost
torch
transformers
sentence-transformers
datasets
accelerate
evaluate

# NLP ES
spacy
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.7.0/es_core_news_md-3.7.0-py3-none-any.whl
nltk

# Explainability (puedes quitarlo si no lo usas de momento)
shap

# MLOps (opcionales, pero útiles)
mlflow
dvc
hydra-core

# API/UI
flask
uvicorn
streamlit

# Viz
matplotlib

# Dev
pyyaml
pytest
flake8
black
isort
